+++ packages/core/src/impl-store.ts
@@ -2,7 +2,7 @@ import debug from './utils/debug';
 import { Manifest, ManifestData, ImplDeployment } from './manifest';
 import { EthereumProvider, isDevelopmentNetwork } from './provider';
 import { Deployment, InvalidDeployment, resumeOrDeploy, waitAndValidateDeployment } from './deployment';
-import type { Version } from './version';
+import { Version } from './version';
 import assert from 'assert';
 import { DeployOpts } from '.';
 
@@ -12,9 +12,10 @@ interface ManifestLens<T> {
   (data: ManifestData): ManifestField<T>;
 }
 
-interface ManifestField<T> {
+export interface ManifestField<T> {
   get(): T | undefined;
   set(value: T | undefined): void;
+  merge?(value: T | undefined): void;
 }
 
 /**
@@ -24,6 +25,7 @@ interface ManifestField<T> {
  * @param provider the Ethereum provider
  * @param deploy the deploy function
  * @param opts options containing the timeout and pollingInterval parameters. If undefined, assumes the timeout is not configurable and will not mention those parameters in the error message for TransactionMinedTimeout.
+ * @param merge if true, adds a deployment to existing deployment by merging their addresses. Defaults to false.
  * @returns the deployment address
  * @throws {InvalidDeployment} if the deployment is invalid
  * @throws {TransactionMinedTimeout} if the transaction was not confirmed within the timeout period
@@ -33,6 +35,7 @@ async function fetchOrDeployGeneric<T extends Deployment>(
   provider: EthereumProvider,
   deploy: () => Promise<T>,
   opts?: DeployOpts,
+  merge?: boolean,
 ): Promise<string> {
   const manifest = await Manifest.forNetwork(provider);
 
@@ -41,14 +44,24 @@ async function fetchOrDeployGeneric<T extends Deployment>(
       debug('fetching deployment of', lens.description);
       const data = await manifest.read();
       const deployment = lens(data);
-      const stored = deployment.get();
-      if (stored === undefined) {
-        debug('deployment of', lens.description, 'not found');
+      if (merge && !deployment.merge) {
+        throw new Error(
+          'fetchOrDeployGeneric was called with merge set to true but the deployment lens does not have a merge function',
+        );
       }
-      const updated = await resumeOrDeploy(provider, stored, deploy);
+
+      const stored = deployment.get();
+      const updated = await resumeOrDeploy(provider, stored, deploy, lens.type, opts, deployment, merge);
       if (updated !== stored) {
-        await checkForAddressClash(provider, data, updated);
-        deployment.set(updated);
+        if (merge && deployment.merge) {
+          // only check primary addresses for clashes, since the address could already exist in an allAddresses field
+          // but the above updated and stored objects are different instances representing the same entry
+          await checkForAddressClash(provider, data, updated, false);
+          deployment.merge(updated);
+        } else {
+          await checkForAddressClash(provider, data, updated, true);
+          deployment.set(updated);
+        }
         await manifest.write(data);
       }
       return updated;
@@ -77,20 +90,56 @@ async function fetchOrDeployGeneric<T extends Deployment>(
   }
 }
 
+/**
+ * Deletes the deployment by setting it to undefined.
+ * Should only be used during a manifest run.
+ */
+export function deleteDeployment(deployment: ManifestField<Deployment>) {
+  deployment.set(undefined);
+}
+
 export async function fetchOrDeploy(
   version: Version,
   provider: EthereumProvider,
   deploy: () => Promise<ImplDeployment>,
   opts?: DeployOpts,
+  merge?: boolean,
 ): Promise<string> {
-  return fetchOrDeployGeneric(implLens(version.linkedWithoutMetadata), provider, deploy, opts);
+  return fetchOrDeployGeneric(implLens(version.linkedWithoutMetadata), provider, deploy, opts, merge);
 }
 
 const implLens = (versionWithoutMetadata: string) =>
   lens(`implementation ${versionWithoutMetadata}`, 'implementation', data => ({
     get: () => data.impls[versionWithoutMetadata],
     set: (value?: ImplDeployment) => (data.impls[versionWithoutMetadata] = value),
+    merge: (value?: ImplDeployment) => {
+      const existing = data.impls[versionWithoutMetadata];
+      if (existing !== undefined && value !== undefined) {
+        const { address, allAddresses } = mergeAddresses(existing, value);
+        data.impls[versionWithoutMetadata] = { ...existing, address, allAddresses };
+      } else {
+        data.impls[versionWithoutMetadata] = value;
+      }
+    },
   }));
+
+/**
+ * Merge the addresses in the deployments and returns them.
+ *
+ * @param existing existing deployment
+ * @param value deployment to add
+ */
+export function mergeAddresses(existing: ImplDeployment, value: ImplDeployment) {
+  const merged = new Set<string>();
+
+  merged.add(existing.address);
+  merged.add(value.address);
+
+  existing.allAddresses?.forEach(item => merged.add(item));
+  value.allAddresses?.forEach(item => merged.add(item));
+
+  return { address: existing.address, allAddresses: Array.from(merged) };
+}
 
 export async function fetchOrDeployAdmin(
   provider: EthereumProvider,
@@ -113,8 +162,9 @@ async function checkForAddressClash(
   provider: EthereumProvider,
   data: ManifestData,
   updated: Deployment,
+  checkAllAddresses: boolean,
 ): Promise<void> {
-  const clash = lookupDeployment(data, updated.address);
+  const clash = lookupDeployment(data, updated.address, checkAllAddresses);
   if (clash !== undefined) {
     if (await isDevelopmentNetwork(provider)) {
       debug('deleting a previous deployment at address', updated.address);
@@ -129,13 +179,20 @@ async function checkForAddressClash(
   }
 }
 
-function lookupDeployment(data: ManifestData, address: string): ManifestField<Deployment> | undefined {
+function lookupDeployment(
+  data: ManifestData,
+  address: string,
+  checkAllAddresses: boolean,
+): ManifestField<Deployment> | undefined {
   if (data.admin?.address === address) {
     return adminLens(data);
   }
 
   for (const versionWithoutMetadata in data.impls) {
-    if (data.impls[versionWithoutMetadata]?.address === address) {
+    if (
+      data.impls[versionWithoutMetadata]?.address === address ||
+      (checkAllAddresses && data.impls[versionWithoutMetadata]?.allAddresses?.includes(address))
+    ) {
       return implLens(versionWithoutMetadata)(data);
     }
   }