+++ packages/plugin-hardhat/src/index.ts
@@ -13,6 +13,7 @@ import type { UpgradeFunction } from './upgrade-proxy';
 import type { DeployBeaconFunction } from './deploy-beacon';
 import type { DeployBeaconProxyFunction } from './deploy-beacon-proxy';
 import type { UpgradeBeaconFunction } from './upgrade-beacon';
+import type { ForceImportFunction } from './force-import';
 import type { ChangeAdminFunction, TransferProxyAdminOwnershipFunction, GetInstanceFunction } from './admin';
 
 export interface HardhatUpgrades {
@@ -22,6 +23,7 @@ export interface HardhatUpgrades {
   deployBeacon: DeployBeaconFunction;
   deployBeaconProxy: DeployBeaconProxyFunction;
   upgradeBeacon: UpgradeBeaconFunction;
+  forceImport: ForceImportFunction;
   silenceWarnings: typeof silenceWarnings;
   admin: {
     getInstance: GetInstanceFunction;
@@ -89,6 +91,7 @@ extendEnvironment(hre => {
     const { makeDeployBeacon } = require('./deploy-beacon');
     const { makeDeployBeaconProxy } = require('./deploy-beacon-proxy');
     const { makeUpgradeBeacon } = require('./upgrade-beacon');
+    const { makeForceImport } = require('./force-import');
     const { makeChangeProxyAdmin, makeTransferProxyAdminOwnership, makeGetInstanceFunction } = require('./admin');
 
     return {
@@ -99,6 +102,7 @@ extendEnvironment(hre => {
       deployBeacon: makeDeployBeacon(hre),
       deployBeaconProxy: makeDeployBeaconProxy(hre),
       upgradeBeacon: makeUpgradeBeacon(hre),
+      forceImport: makeForceImport(hre),
       admin: {
         getInstance: makeGetInstanceFunction(hre),
         changeProxyAdmin: makeChangeProxyAdmin(hre),