+++ packages/plugin-hardhat/src/utils/options.ts
@@ -20,4 +20,5 @@ export interface DeployProxyOptions extends Options {
 
 export interface UpgradeProxyOptions extends Options {
   call?: { fn: string; args?: unknown[] } | string;
+  unsafeSkipStorageCheck?: boolean;
 }